test_collection/relevance_judgements/umbrela/pyproject.toml
[project]
name = "umbrela"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.8,<4",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "retry>=0.9.2",
    "scikit-learn>=1.8.0",
    "setuptools>=80.9.0",
    "tqdm>=4.67.1",
    "typing-extensions>=4.15.0",
]
